refactor: move checkpoint-hint identity logic and tests to last_checkpoint_hint
Add LastCheckpointHint::applies_to(version, checkpoint_parts) -- the identity gate formerly inline in LogSegment::checkpoint_hint(), which now delegates to it. This lets the _last_checkpoint hint tests live in last_checkpoint_hint: the gate is tested directly via applies_to (no LogSegment), and the real-table content test moves over. Lift create_log_path helpers to crate::utils::test_utils so both modules share them. The synthetic LogSegment-based accessor tests are dropped except one checkpoint_sidecars empty-vs-absent contract test, which needs a LogSegment and so stays in log_segment.
